requirements.txt
Flask==3.0.2
mysql-connector-python==8.3.0
google-generativeai==0.3.2
pdfkit==1.0.0
python-dotenv==1.0.1
requests==2.31.0
pandas
